Fast Bugfix for new needs required by for ... in ... construct.
Previous requirement on iterators was operator!= and operator++
Now is operator< and operator++, for compatibility with OpenMP.

+++ pythran/pythonic++/core/intrinsics.h
@@ -28,7 +28,7 @@ namespace pythonic {
     /* bin */
     template<class T>
         std::string bin(T const &v) {
-            size_t i = 1 << (8*sizeof(T)-1);
+            size_t i = (size_t)1 << (8*sizeof(T)-1);
             while(i and not (v&i)) i>>=1;
             if(not i) return "0b0";
             else {
